+++ include/caffe/layers/loss_layer.hpp
@@ -10,7 +10,7 @@
 namespace caffe
 {
 
-  const float kLOG_THRESHOLD = 1e-20;
+  const double kLOG_THRESHOLD = 1e-20;
 
   /**
    * @brief An interface for Layer%s that take two Blob%s as input -- usually
